% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_name.R
\name{check_pkg_name}
\alias{check_pkg_name}
\title{Check Naming Convention rules}
\usage{
check_pkg_name(pkg_name)
}
\arguments{
\item{pkg_name}{character name of a package.}
}
\description{
Helps package creation function to reject names that do not comply.
}
