% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_pkg.R
\name{create_data_pkg}
\alias{create_data_pkg}
\title{Create a New openwashdata Dataset R Package.}
\usage{
create_data_pkg(
  pkg_name,
  user.name = NULL,
  user.email = NULL,
  location = NULL,
  open = FALSE
)
}
\arguments{
\item{pkg_name}{The name of the package to create.}

\item{user.name}{The name of the package author.}

\item{user.email}{The email address of the package author.}

\item{location}{The location where the package should be created.}

\item{open}{A logical value indicating whether to open the package directory in a new window.}
}
\description{
This function creates a new data R package with the specified name and author information.
It automatically creates a git repository and remote git data repository within the openwashdata
GitHub organization. Note that the functions requires a GitHub account that
has write access to the openwashdata GitHub organization as well as up and running RSA
key pair authentication with GitHub set up. Personal access tokens will be
created in the process though.
}
\examples{
create_data_pkg("my_package", "John Doe", "johndoe@example.com")
}
