% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dataset_doc.R
\name{generate_dataset_documentation}
\alias{generate_dataset_documentation}
\title{Create Documentation from Dictionary Bootstrap File}
\usage{
generate_dataset_documentation(
  dataset_name,
  dataset_title,
  dict_path = NULL,
  output_path = NULL,
  split_by_dataset = TRUE
)
}
\arguments{
\item{dataset_name}{character vector names of the dataset.}

\item{dataset_title}{character vector short, one line description of the dataset.}

\item{dict_path}{character path to the dictionary .csv files. Defaults to NULL (bootstrap folder).}

\item{output_path}{character path for the output. Defaults to NULL, using the R/ folder.}
}
\description{
Use a dictionary file to create a roxygen2 style R data package documentation entry.
}
