% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Copy Workflow Template}
\usage{
init(target = file.path(getwd(), "bootstrap_pkg.R"))
}
\description{
Copies workflow template file into the current working directory
and opens the file if RStudio is the editor.
}
