% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_data_pkg}
\alias{init_data_pkg}
\title{Initialize Data Package}
\usage{
init_data_pkg(pkg_dir_name, path = NULL)
}
\arguments{
\item{pkg_dir_name}{character folder name of the package}

\item{path}{character parent directory of the package folder.}
}
\description{
Initialize Data Package
}
