% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_pkg.R
\name{init_git}
\alias{init_git}
\title{Add git to the Data Package Skeleton}
\usage{
init_git(user.name = NULL, user.email = NULL)
}
\arguments{
\item{user.name}{character user name. Defaults to NULL, using global defaults.}

\item{user.email}{character user name. Defaults to NULL, using global defaults.}
}
\description{
Initialize a git repository convenience wrapper
}
