% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openwashdata-package.R
\docType{package}
\name{openwashdata-package}
\alias{openwashdata}
\alias{openwashdata-package}
\title{openwashdata: This package provides support functions for creating openwashdata R data packages}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lars Schöbitz \email{lschoebitz@ethz.ch} (\href{https://orcid.org/0000-0003-2196-5015}{ORCID})

}
\keyword{internal}
