% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roles.R
\name{parse_roles}
\alias{parse_roles}
\title{Parse Roles from DESCRIPTION}
\usage{
parse_roles(desc_path = "DESCRIPTION", roles = c("aut", "cre", "ctb", "fnd"))
}
\arguments{
\item{desc_path}{character path to DESCRIPTION file.}

\item{roles}{character vector of roles to look for.}
}
\description{
Parse Roles from DESCRIPTION
}
