% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_pkg.R
\name{push_package}
\alias{push_package}
\title{Push Data Package to openwashdata on GitHub}
\usage{
push_package()
}
\description{
Pushes and a data package to the opendata GitHub repository. Note that you
will need an openwasdata personal access token to do so.
}
