% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_authors.R
\name{use_authors_from_csv}
\alias{use_authors_from_csv}
\title{Read Authors from a CSV File and Add Authors to DESCRIPTION}
\usage{
use_authors_from_csv(f = "data-raw/authors.csv", quiet = FALSE)
}
\arguments{
\item{f}{character file location of the authors .csv file}

\item{quiet}{logical should the function be quiet about the number of authors
found in the csv file? Defaults to FALSE.}
}
\description{
Use a specifically formatted .csv file to read in authors and their properties
and add them to the current project's DESCRIPTION file. This is implies that
the current project is an openwashdata data package skeleton.
}
