% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_authors.R
\name{use_authors_skeleton}
\alias{use_authors_skeleton}
\title{Create a Simple Skeleton to Conveniently Gather Author Information}
\usage{
use_authors_skeleton(dest = NULL)
}
\arguments{
\item{dest}{character full path to authors skeleton file. Defaults to NULL.}
}
\description{
Create a Simple Skeleton to Conveniently Gather Author Information
}
