% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dictionary.R
\name{use_dictionary}
\alias{use_dictionary}
\title{Extract Dictionary from Data Folder to Create a Pre-Filled Skeleton}
\usage{
use_dictionary(
  data_location = NULL,
  skeleton_dest = "bootstrap/dictionary.csv",
  data_file_pattern = ".Rda|.rda",
  ignore_pattern = "codebook.Rda",
  recursive = TRUE
)
}
\arguments{
\item{data_location}{character folder location of the raw_data folder.}

\item{skeleton_dest}{character folder destination of the skeleton file.}

\item{data_file_pattern}{character which data files should be parsed.
For now, only .Rda is available.}

\item{ignore_pattern}{character regular expression of file names to ignore}

\item{recursive}{boolean should this function iterate through
subfolders of data-raw? Defaults to TRUE.}
}
\description{
Extract Dictionary from Data Folder to Create a Pre-Filled Skeleton
}
